Add multistage docker builds
There is now  a build for local and a build for production.
Currently, the NODE_ENV variable has not been set for either
build stage; however, the production stage will not install
dev-dependencies. Nodemon has been installed as a dev-dependency,
and with the volume-mount in the web service of docker-compose,
changes are propogated to the container for ease of local
development.

The default build is production. The base
image uses Node's LTS. The Dockerfile, docker-compose.yaml, and
node_modules has been added to .dockerignore.